# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.Class:
    properties:
      agencyId:
        type: integer
      createdAt:
        type: string
      grade:
        description: |-
          Grade type:
          * buds - Children who is 3 yo.
          * seed - Children who is 4 yo.
          * leaf - Children who is 5 yo.
          * toddler - Children who are lower than 2 yo.
        enum:
        - buds
        - seed
        - leaf
        - ' toddler'
        type: string
      id:
        type: integer
      name:
        type: string
      teacherId:
        type: string
      updatedAt:
        type: string
    type: object
  entities.Student:
    properties:
      agencyId:
        type: integer
      birthPlace:
        type: string
      class:
        $ref: '#/definitions/entities.Class'
      classId:
        type: integer
      createdAt:
        type: string
      dob:
        type: string
      enrolledAt:
        type: string
      ethnic:
        type: string
      father_birth_place:
        type: string
      father_dob:
        type: string
      father_name:
        type: string
      father_occupation:
        type: string
      father_phone_number:
        type: string
      firstName:
        type: string
      gender:
        type: boolean
      id:
        type: integer
      land_lord:
        type: string
      lastName:
        type: string
      mother_birth_place:
        type: string
      mother_dob:
        type: string
      mother_name:
        type: string
      mother_occupation:
        type: string
      mother_phone_number:
        type: string
      parent_res_registration:
        type: string
      parent_roi:
        type: string
      parent_zalo:
        type: string
      permanentAddressCommune:
        type: string
      permanentAddressDistrict:
        type: string
      permanentAddressProvince:
        type: string
      tempAddress:
        type: string
      updatedAt:
        type: string
    type: object
...
  students.FindStudentResp:
    properties:
      data:
        items:
          $ref: '#/definitions/entities.Student'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
    type: object
  students.Student:
    properties:
      agencyId:
        type: integer
      birthPlace:
        type: string
      class:
        $ref: '#/definitions/entities.Class'
      classId:
        type: integer
      createdAt:
        type: string
      dob:
        type: string
      enrolledAt:
        type: string
      ethnic:
        type: string
      father_birth_place:
        type: string
      father_dob:
        type: string
      father_name:
        type: string
      father_occupation:
        type: string
      father_phone_number:
        type: string
      firstName:
        type: string
      gender:
        type: boolean
      id:
        type: integer
      land_lord:
        type: string
      lastName:
        type: string
      mother_birth_place:
        type: string
      mother_dob:
        type: string
      mother_name:
        type: string
      mother_occupation:
        type: string
      mother_phone_number:
        type: string
      parent_res_registration:
        type: string
      parent_roi:
        type: string
      parent_zalo:
        type: string
      permanentAddressCommune:
        type: string
      permanentAddressDistrict:
        type: string
      permanentAddressProvince:
        type: string
      tempAddress:
        type: string
      updatedAt:
        type: string
    type: object
  students.WriteStudentRequest:
    properties:
      agencyId:
        type: integer
      birthPlace:
        type: string
      classId:
        type: integer
      dob:
        type: string
      enrolledAt:
        type: string
      ethnic:
        type: string
      father_birth_place:
        type: string
      father_dob:
        type: string
      father_name:
        type: string
      father_occupation:
        type: string
      father_phone_number:
        type: string
      firstName:
        type: string
      gender:
        type: boolean
      land_lord:
        type: string
      lastName:
        type: string
      mother_birth_place:
        type: string
      mother_dob:
        type: string
      mother_name:
        type: string
      mother_occupation:
        type: string
      mother_phone_number:
        type: string
      parent_res_registration:
        type: string
      parent_roi:
        type: string
      parent_zalo:
        type: string
      permanentAddressCommune:
        type: string
      permanentAddressDistrict:
        type: string
      permanentAddressProvince:
        type: string
      tempAddress:
        type: string
    type: object
  users.FindUserResp:
    properties:
      data:
        items:
          $ref: '#/definitions/users.User'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
    type: object
  users.User:
    properties:
      agencyId:
        type: integer
      auth0_user_id:
        type: string
      connection:
        type: string
      createdAt:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      password:
        description: Plaintext password for input
        type: string
      phone_number:
        type: string
      updatedAt:
        type: string
      verify_email:
        type: boolean
    type: object
  users.UserInput:
    properties:
      agencyId:
        type: integer
      auth0_user_id:
        type: string
      connection:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      password:
        description: Plaintext password for input
        type: string
      phone_number:
        type: string
      verify_email:
        type: boolean
    type: object
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /sign-up:
    post:
      consumes:
      - application/json
      description: Insert user by sign up
      parameters:
      - description: Create user sign up body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/users.UserInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create user by sign up api
      tags:
      - SignUpInTheUserRoute
  /students:
    delete:
      consumes:
      - application/json
      description: Delete student
      parameters:
      - collectionFormat: csv
        description: Student IDs
        in: query
        items:
          type: integer
        name: ids
        required: true
        type: array
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete student api
      tags:
      - Student
    get:
      consumes:
      - application/json
      description: Get list student
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: pageSize
        type: integer
      - default: desc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/students.FindStudentResp'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get list student details api
      tags:
      - Student
    post:
      consumes:
      - application/json
      description: Insert student
      parameters:
      - description: Create student body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/students.WriteStudentRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/students.Student'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create student api
      tags:
      - Student
  /students/{id}:
    get:
      consumes:
      - application/json
      description: Get one student
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/students.Student'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get student details api
      tags:
      - Student
    put:
      consumes:
      - application/json
      description: Update student
      parameters:
      - description: Update student body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/students.WriteStudentRequest'
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/students.Student'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update student api
      tags:
      - Student
  /users:
    get:
      consumes:
      - application/json
      description: Get list user
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: pageSize
        type: integer
      - default: desc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Email term
        in: query
        name: email
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.FindUserResp'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get list user details api
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Insert user
      parameters:
      - description: Create user body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/users.UserInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create user api
      tags:
      - User
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get one user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get user details api
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: Update user body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/users.UserInput'
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update user api
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    description: Apply "bearer " before token in authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"